+++ src/main/java/roomescape/domain/member/Member.java
@@ -1,4 +1,4 @@
-package roomescape.domain;
+package roomescape.domain.member;
 
 import jakarta.persistence.*;
