+++ src/main/java/roomescape/domain/repository/PaymentRepository.java
@@ -2,7 +2,7 @@ package roomescape.domain.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import roomescape.domain.Payment;
+import roomescape.domain.payment.Payment;
 
 import java.util.Optional;
