+++ src/main/java/roomescape/domain/repository/ReservationTimeRepository.java
@@ -2,7 +2,7 @@ package roomescape.domain.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import roomescape.domain.ReservationTime;
+import roomescape.domain.reservation.ReservationTime;
 
 import java.time.LocalTime;
