+++ src/main/java/roomescape/domain/repository/ReservationWaitingRepository.java
@@ -3,7 +3,12 @@ package roomescape.domain.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
-import roomescape.domain.*;
+import roomescape.domain.member.Member;
+import roomescape.domain.reservation.ReservationDate;
+import roomescape.domain.reservation.ReservationTime;
+import roomescape.domain.reservation.ReservationWaiting;
+import roomescape.domain.reservation.ReservationWaitingWithRank;
+import roomescape.domain.reservation.theme.Theme;
 
 import java.util.List;
 import java.util.Optional;
