+++ src/main/java/roomescape/domain/repository/ThemeRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
-import roomescape.domain.Theme;
+import roomescape.domain.reservation.theme.Theme;
 
 import java.time.LocalDate;
 import java.util.List;
