+++ src/main/java/roomescape/domain/reservation/Reservation.java
@@ -1,6 +1,9 @@
-package roomescape.domain;
+package roomescape.domain.reservation;
 
 import jakarta.persistence.*;
+import roomescape.domain.member.Member;
+import roomescape.domain.payment.Payment;
+import roomescape.domain.reservation.theme.Theme;
 
 @Entity
 public class Reservation {
