+++ src/main/java/roomescape/domain/reservation/ReservationDate.java
@@ -1,4 +1,4 @@
-package roomescape.domain;
+package roomescape.domain.reservation;
 
 import jakarta.persistence.Embeddable;
