+++ src/main/java/roomescape/service/PaymentService.java
@@ -2,8 +2,8 @@ package roomescape.service;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import roomescape.domain.Payment;
-import roomescape.domain.Reservation;
+import roomescape.domain.payment.Payment;
+import roomescape.domain.reservation.Reservation;
 import roomescape.domain.repository.PaymentRepository;
 import roomescape.domain.repository.ReservationRepository;
 import roomescape.infrastructure.payment.PaymentManager;
