+++ src/main/java/roomescape/service/response/ReservationTimeDto.java
@@ -2,7 +2,7 @@ package roomescape.service.response;
 
 import java.time.LocalTime;
 
-import roomescape.domain.ReservationTime;
+import roomescape.domain.reservation.ReservationTime;
 
 public record ReservationTimeDto(Long id, LocalTime startAt) {
