+++ src/test/java/roomescape/controller/AdminPageControllerTest.java
@@ -13,7 +13,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import roomescape.domain.MemberRole;
+import roomescape.domain.member.MemberRole;
 
 class AdminPageControllerTest extends ControllerTest {
     @BeforeEach
