+++ src/test/java/roomescape/controller/MemberAuthControllerTest.java
@@ -14,7 +14,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
 import org.junit.jupiter.params.provider.ValueSource;
-import roomescape.domain.MemberRole;
+import roomescape.domain.member.MemberRole;
 import roomescape.web.controller.request.MemberSignUpRequest;
 import roomescape.web.controller.request.TokenRequest;
