+++ src/test/java/roomescape/controller/PaymentControllerTest.java
@@ -7,8 +7,12 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import roomescape.controller.doc.DocumentFilter;
-import roomescape.domain.*;
+import roomescape.domain.member.Member;
 import roomescape.domain.repository.*;
+import roomescape.domain.reservation.Reservation;
+import roomescape.domain.reservation.ReservationDate;
+import roomescape.domain.reservation.ReservationTime;
+import roomescape.domain.reservation.theme.Theme;
 import roomescape.infrastructure.auth.JwtProvider;
 import roomescape.infrastructure.payment.PaymentManager;
 import roomescape.service.request.PaymentApproveDto;
