+++ src/test/java/roomescape/domain/MemberRoleTest.java
@@ -7,6 +7,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
 import org.junit.jupiter.params.provider.ValueSource;
+import roomescape.domain.member.MemberRole;
 
 class MemberRoleTest {
     @DisplayName("생성 테스트")
