+++ src/test/java/roomescape/domain/MemberTest.java
@@ -8,6 +8,8 @@ import static roomescape.Fixture.VALID_USER_PASSWORD;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import roomescape.domain.member.Member;
+import roomescape.domain.member.MemberRole;
 
 class MemberTest {
