+++ src/test/java/roomescape/domain/ReservationDateTest.java
@@ -7,6 +7,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import roomescape.domain.reservation.ReservationDate;
 
 public class ReservationDateTest {
