+++ src/test/java/roomescape/domain/ReservationTest.java
@@ -10,6 +10,8 @@ import static roomescape.Fixture.VALID_THEME;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import roomescape.domain.reservation.Reservation;
+import roomescape.domain.reservation.ReservationDate;
 
 class ReservationTest {
