+++ src/test/java/roomescape/service/ReservationWaitingServiceTest.java
@@ -6,8 +6,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.jdbc.Sql;
 import org.springframework.transaction.annotation.Transactional;
-import roomescape.domain.*;
+import roomescape.domain.member.*;
 import roomescape.domain.repository.*;
+import roomescape.domain.reservation.*;
+import roomescape.domain.reservation.theme.Theme;
 import roomescape.service.request.ReservationWaitingSaveDto;
 import roomescape.service.response.ReservationWaitingDto;
 import roomescape.service.response.ReservationWaitingWithRankDto;
